MAXPLAYERS is a usize, and player numbers are PlayerId or slot indices
MAXPLAYERS was an i32 that every use immediately converted: 26 casts to
usize, 3 to u32, ~20 `for i in 0..MAXPLAYERS` loops whose variable then
only indexed. Flipping it to usize exposed 33 mismatches, all of the
same kind: a player number kept as a plain i32 that was passed to a
function or compared with another i32. Those are retyped instead of
casted:

PlayerId (the typed slot that already existed):
  do_reborn, check_spot, death_match_spawn_player, player_finish_level,
  player_reborn, player_quit_game; their loops are `for player in
  PlayerId::all()` (g_ticker, do_completed, P_SetupLevel, run_tic).
  mobj_t::lastlook is a PlayerId; look_for_players uses next_wrapping()
  and a new previous_wrapping() instead of `(x +/- 1) & 3`. Loaded
  savegames reduce the stored value mod MAXPLAYERS (vanilla only ever
  stores 0..3, so this only matters for a corrupt file).
usize slot indices (intermission screen, chat):
  WbStartStruct::pnum, WiStuffState::me, plyr_index, frag_sum,
  bind_chat_controls(num_players), NetConnectData::max_players.

PlayerId::all(), next_wrapping() and PerPlayer<T> hard-coded 4; they use
MAXPLAYERS now. So do the arrays that are one-entry-per-player (frags,
playeringame, playerstarts, turbodetected, consistancy, chat_dest,
w_inputbuffer, PLAYER_NAMES, and the intermission screen's per-player
tables). Left as 4: numanims (episodes), frameskip (`frameon & 3`), the
ammo arrays, BBox.

Behaviour is unchanged. Not covered by the golden tests: deathmatch and
netgame paths (deathmatch stats screens, chat, player_quit_game, netgame
respawn placement); those are type-checked only.

Verified: cargo test --release (57 pass, goldens unchanged), clippy and
rustfmt clean, workspace builds.

+++ rust-doomgeneric/engine/src/d_main.rs
@@ -407,7 +407,7 @@ pub fn bind_variables(m_config: &mut MConfigState, m_controls: &mut MControlsSta
     bind_weapon_controls(m_config);
     bind_map_controls(m_config);
     bind_menu_controls(m_config);
-    bind_chat_controls(m_config, MAXPLAYERS as u32);
+    bind_chat_controls(m_config, MAXPLAYERS);
     m_controls.key_multi_msgplayer[0] = HUSTR_KEYGREEN;
     m_controls.key_multi_msgplayer[1] = HUSTR_KEYINDIGO;
     m_controls.key_multi_msgplayer[2] = HUSTR_KEYBROWN;
