+++ rust-doomgeneric/engine/src/doomdef.rs
@@ -1,5 +1,5 @@
 pub type Pixel = u32;
-pub const MAXPLAYERS: i32 = 4;
+pub const MAXPLAYERS: usize = 4;
 pub const TICRATE: i32 = 35;
 pub const SCREENWIDTH: i32 = 320;
 pub const SCREENHEIGHT: i32 = 200;
